requirements.txt
fastapi
uvicorn
requests
python-dotenv
langchain
openai
aiohttp
langchain_community
azure-identity
azure-keyvault-secrets